pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-content-auditor"
version = "0.1.0"
description = "AI-powered content auditor for SEO, accessibility, and brand tone analysis"
authors = [{name = "ECU Team"}]
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.0",
    "ollama>=0.1.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
sheets = [
    "google-api-python-client>=2.0.0",
]

[project.scripts]
audit-content = "main:app"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []
